+++ R/utils.R
@@ -150,7 +150,7 @@
 #'   \item \emph{\code{corr_wmean}}: For corrected weighted mean.
 #' }
 #' @param subsample \strong{\code{\link[base]{numeric}}} | Number of cells to randomly select from the Seurat object to enhance performance. Selecting NA will disable this but might lead to function breaks if the sample size is too large.
-#' @param cluster_cells \strong{\code{\link[base]{logical}}} | Whether to cluster internally the cells in each group. If the group is too large, this will disable automatically.
+#' @param cluster \strong{\code{\link[base]{logical}}} | Whether to perform clustering of rows and columns.
 #' @param subsample \strong{\code{\link[base]{numeric}}} | Number of cells to subset for the analysis. NA will use all. Cells are selected at random.
 #' @param plot.title.face,plot.subtitle.face,plot.caption.face,axis.title.face,axis.text.face,legend.title.face,legend.text.face \strong{\code{\link[base]{character}}} | Controls the style of the font for the corresponding theme element.  One of:
 #' \itemize{
@@ -160,6 +160,8 @@
 #'   \item \emph{\code{bold.italic}}: For text both in itallic and bold.
 #' }
 #' @param flavor \strong{\code{\link[base]{character}}} | One of: Seurat, UCell. Compute the enrichment scores using \link[Seurat]{AddModuleScore} or \link[UCell]{AddModuleScore_UCell}.
+#' @param features.order \strong{\code{\link[base]{character}}} | Should the gene sets be ordered in a specific way? Provide it as a vector of characters with the same names as the names of the gene sets.
+#' @param groups.order \strong{\code{\link[SCpubr]{named_list}}} | Should the groups in theheatmaps be ordered in a specific way? Provide it as a named list (as many lists as values in \strong{\code{group.by}}) with the order for each of the elements in the groups.
 #' @usage NULL
 #' @return Nothing. This is a mock function.
 #' @keywords internal
@@ -275,7 +277,9 @@ doc_function <- function(sample,
                          axis.text.face,
                          legend.title.face,
                          legend.text.face,
-                         flavor){}
+                         flavor,
+                         features.order,
+                         groups.order){}
 
 #' Named vector.
 #'