+++ tests/testthat/test-do_SCEnrichmentHeatmap.R
@@ -18,6 +18,22 @@ if (isFALSE(dep_check[["do_SCEnrichmentHeatmap"]])){
   
   testthat::test_that("do_SCEnrichmentHeatmap: PASS - default", {
     testthat::skip_on_cran()
+    
+    
+    genes <- list("A" = rownames(sample)[1:5],
+                  "B" = rownames(sample)[6:10],
+                  "C" = rownames(sample)[11:15])
+    
+    p <- SCpubr::do_SCEnrichmentHeatmap(sample = sample,
+                                        input_gene_list = genes,
+                                        flavor = "Seurat",
+                                        assay = "SCT",
+                                        nbin = 1,
+                                        ctrl = 5,
+                                        cluster = FALSE,
+                                        features.order = c("B", "C", "A"))
+    testthat::expect_type(p, "list")
+    
     genes <- list("A_A" = rownames(sample)[1:5],
                   "B_A" = rownames(sample)[6:10],
                   "C_A" = rownames(sample)[11:15])
@@ -70,7 +86,7 @@ if (isFALSE(dep_check[["do_SCEnrichmentHeatmap"]])){
                                         assay = "SCT",
                                         nbin = 1,
                                         ctrl = 5,
-                                        cluster_cells = TRUE)
+                                        cluster = TRUE)
     testthat::expect_type(p, "list")
     
     p <- SCpubr::do_SCEnrichmentHeatmap(sample = sample,
@@ -79,7 +95,7 @@ if (isFALSE(dep_check[["do_SCEnrichmentHeatmap"]])){
                                         assay = "SCT",
                                         nbin = 1,
                                         ctrl = 5,
-                                        cluster_cells = FALSE)
+                                        cluster = FALSE)
     testthat::expect_type(p, "list")
     
     p <- SCpubr::do_SCEnrichmentHeatmap(sample = sample,